optimize: optimize code logic and fixed error of type convert

+++ src/main/java/org/conch/http/JSONData.java
@@ -478,10 +478,10 @@ public final class JSONData {
     public static JSONObject forkBlock(Block block) {
         JSONObject json = new JSONObject();
         json.put("block", block.getStringId());
-        json.put("height", block.getHeight());
+        json.put("height", (long) block.getHeight());
         json.put("generatorRS", Account.rsAccount(block.getGeneratorId()));
-        json.put("timestamp", block.getTimestamp());
-        json.put("version", block.getVersion());
+        json.put("timestamp", (long) block.getTimestamp());
+        json.put("version", (long) block.getVersion());
         json.put("cumulativeDifficulty", block.getCumulativeDifficulty().toString());
 //        if (block.getNextBlockId() != 0) {
 //            json.put("nextBlock", Long.toUnsignedString(block.getNextBlockId()));
